Altered Registration of Token
There is no need to grab a runner token before the CloudFormation initialization using the POST request. You can register the runner via command using the Registration Token from GitLab UI. Even though the configuration isn't as simple as it would be using the config.toml, it can be done. 

Grabbing the token from the POST request (as described in README), does not pass information like name or description of the runner. This method registers the runner with the appropriate attributes.

I couldn't find a list of the variables in GitLab's documentation, but there is a small summary here: http://www.oznetnerd.com/ci-cd-setting-up-gitlab-runners-on-aws-using-coreos-terraform/